[FIFO] move characterization into HLSCustomOp member fxns
